newNavbar removed and navbar code cleaned
